% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTemps.R
\name{convertTemps}
\alias{convertTemps}
\title{Convert contiguous U.S. (CONUS) temperatures to global temperatures or vice versa}
\usage{
convertTemps(temps, from)
}
\arguments{
\item{temps}{A numeric vector of CONUS or global temperatures in degrees Celsius.}

\item{from}{A character string (one of  \code{c("conus", "global")}), indicating whether users are converting from CONUS to global temperatures (\code{from="conus"}) or from global to CONUS (\code{from="global"}).

Temperatures for the contiguous U.S. (CONUS) in degrees Celsius are converted to global temperatures (\code{convertTemps(temps, from="conus")}) in degrees Celsius by multiplying CONUS temperatures by a constant \emph{k} (where \emph{k=1.421}).

Global temperatures in degrees Celsius are converted to CONUS temperatures (\code{convertTemps(temps, from="global")}) in degrees Celsius by multiplying global temperatures by a constant \emph{1/k} (where \emph{k=1.421}).}
}
\value{
Outputs a numeric vector of temperatures in degrees Celsius.
}
\description{
This pre-processing helper function converts a list of warming temperatures in degrees Celsius (\code{temps}) from global to CONUS (\code{from="global"}) or vice versa (\code{from="global"}). The equations for converting between CONUS and global temperatures and back again are described elsewhere in this report.
}
\examples{
convertTemps(1:7, from = "global")
convertTemps(1:7, from = "conus")

}
\references{
Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at \url{https://epa.gov/cira/FrEDI/}.
}
