% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configTests.R
\name{dataInfo_test}
\alias{dataInfo_test}
\title{test_loadData}
\usage{
dataInfo_test(
  dataList = list(),
  outPath = ".",
  csvName = "dataInfo_test",
  save = TRUE,
  return = TRUE
)
}
\arguments{
\item{outPath}{A filepath to where test files are saved}

\item{save}{A TRUE/FALSE option to save files or not}

\item{return}{A TRUE/FALSE option to return the test_table}

\item{data_list}{A list of named tables to be examined}
}
\value{
Returns a table object if "return" option is TRUE
}
\description{
test_loadData
}
