% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrEDI-package.R
\docType{package}
\name{FrEDI-package}
\alias{FrEDI}
\alias{FrEDI-package}
\title{README
FrEDI: The Framework for Evaluating Damages and Impacts}
\description{
\link{FrEDI} is an R package being developed by the U.S. Environmental Protection Agency (EPA). The functions and data provided by this package can be used to estimate climate change impacts for the contiguous United States (CONUS) using the Framework for Evaluating Damages and Impacts (FrEDI), developed as part of EPA's \href{https://epa.gov/cira/}{Climate Change Impacts and Risk Analysis} (CIRA) project. \link{FrEDI} contains R code that implement FrEDI and allow users to project impacts from climate change and sea level rise for a selected set of sectors.
}
\details{
For help getting started with \link{FrEDI}, visit \url{https://usepa.github.io/FrEDI/articles/manual.html}.

For additional package documentation, see \url{https://usepa.github.io/FrEDI/index.html}.

For more information on the Framework and the CIRA project, visit \url{https://epa.gov/cira/}, especially \url{https://epa.gov/cira/FrEDI/}.
}
\section{Overview of Functions}{

The function \code{\link[=run_fredi]{run_fredi()}} provided in this package is the primary function implementing \code{FrEDI} The main inputs to \code{\link[=run_fredi]{run_fredi()}} are climate scenarios (temperature in degrees Celsius, global mean sea level rise in centimeters) and socioeconomic scenarios (gross domestic product, regional population). \code{\link[=run_fredi]{run_fredi()}} projects climate impacts for these scenarios using the temperature binning framework. Users can also specify the levels at which to aggregate the climate impacts.

\link{FrEDI} also contains functions to assist in the pre-processing of input scenarios and the post-processing of outputs.
\itemize{
\item Pre-processing functions include \code{\link[=get_sectorInfo]{get_sectorInfo()}}, \code{\link[=import_inputs]{import_inputs()}}, \code{\link[=convertTemps]{convertTemps()}}, \code{\link[=temps2slr]{temps2slr()}}.
\itemize{
\item \code{\link[=get_sectorInfo]{get_sectorInfo()}} allows users to access a list of sectors within \code{FrEDI} and related sector information.
\item \code{\link[=import_inputs]{import_inputs()}} helps users in importing custom scenarios from user-specified comma-separated value (CSV) files.
\item \code{\link[=convertTemps]{convertTemps()}} helps users to convert between global mean temperature and temperatures for the contiguous United States (CONUS) (both in degrees Celsius).
\item \code{\link[=temps2slr]{temps2slr()}} helps users to estimate global mean sea level rise (GMSL, in centimeters) from global mean temperature in degrees Celsius.
}
}

\code{\link[=aggregate_impacts]{aggregate_impacts()}} is a post-processing helper function that helps users to aggregate and summarize the outputs of \link{FrEDI}. \code{\link[=aggregate_impacts]{aggregate_impacts()}} can be used to calculate national totals, model averages, sum over impact types, and interpolate between multiple impact years (note that \code{\link[=run_fredi]{run_fredi()}} will automatically run \code{\link[=aggregate_impacts]{aggregate_impacts()}} before returning outputs if the \code{aggLevels} argument is not \code{"none"}).

Versions 2.3.0 and above include the \code{FrEDI} Social Vulnerability (SV) module for estimating impacts on socially vulnerable populations for select sectors. \code{\link[=get_sv_sectorInfo]{get_sv_sectorInfo()}} allows users to access a list of sectors within the FrEDI SV module and related sector information. The function \code{\link[=run_fredi_sv]{run_fredi_sv()}} is the main function for the \code{FrEDI} SV module. \code{\link[=run_fredi_sv]{run_fredi_sv()}} is designed to calculate impacts for a single sector at a time for a custom population scenario or one or more custom temperature or sea level rise scenarios. When running \code{\link[=run_fredi_sv]{run_fredi_sv()}}, users have the option to save results to an Excel file that provides basic data visualization for the module outputs. For more information on the data underlying the \code{FrEDI} SV module, visit \url{https://www.epa.gov/cira/social-vulnerability-report/}.

@section Overview of Package Contents:
\link{FrEDI} consists of files in the following directories:
\itemize{
\item \strong{R}. Contains function definitions (files ending in \code{".R"}) and configuration files (ending in \code{".rda"}).
\item \strong{data}. Contains R Data files ending in \code{".rdb"}, \code{".rds"}, and \code{".rdx"}, containing data included with the package.
\item \strong{help} and \strong{html}. Contain documentation for functions available to the user, including function descriptions, lists of arguments and outputs, and examples. See \code{"html/00Index.html"} or the individual R help functions for more information about individual functions.
\item \strong{Meta}. Contains RDS files (ending in \code{".rds"}) with information about the package contents.
\item \strong{extdata}. \strong{extdata/scenarios} Contains four CSV files for users to test the function for importing data. For more information on importing scenarios for use with \code{\link[=run_fredi]{run_fredi()}}, refer to documentation for the function \code{\link[=import_inputs]{import_inputs()}}.
\itemize{
\item \code{"GCAM_scenario.csv"} contains a set of temperature scenarios that can be used with \link{FrEDI}, including the default temperature scenario used by both \code{\link[=run_fredi]{run_fredi()}} and \code{\link[=run_fredi_sv]{run_fredi_sv()}} (labeled \verb{"description="Hector_GCAM_v5.3_ECS_3.0_REF"} in \code{"slr_from_GCAM.csv"} and \code{"ECS_3.0_ref_0"} in the \link{gcamScenarios} dataset)). Also see the documentation for the \link{gcamScenarios} dataset for more information.
\item \code{"State ICLUS Population.csv"} and \code{"Region ICLUS Population.csv"} contain the default population scenarios used by \code{\link[=run_fredi]{run_fredi()}} and \code{\link[=run_fredi_sv]{run_fredi_sv()}}, respectively (see \link{popScenario} and \link{popScenario_sv}).
\item \code{"slr_from_GCAM.csv"} contains global mean sea level rise heights in centimeters (created from the reference temperature scenario).
}
\item \strong{extdata/sv} Contains files used by the \code{FrEDI} SV module to calculate impacts.
}

The \code{FrEDI} package contains a loadable dataset with default results \code{defaultResults}, which contains annual impacts produced by \code{\link[=run_fredi]{run_fredi()}} for the with the default options and default default scenarios (i.e., default temperature, GDP and regional population trajectories). Other loadable datasets provided by \link{FrEDI} are a set of driver scenarios (\link{gcamScenarios}), a state-level population scenario (\link{popScenario}) for use with \code{\link[=run_fredi]{run_fredi()}}, and a region-level \code{popScenario_sv} for use with \code{\link[=run_fredi_sv]{run_fredi_sv()}}. These data can be loaded into the workspace using the \code{data()} function.

Typical use will involve \code{library(FrEDI)} or \code{require(FrEDI)}.
}

\section{Status}{

Disclaimer: All code in this repository is being provided in a "draft" state and has not been reviewed or cleared by U.S. EPA. This status will be updated as models are reviewed.
}

\section{Dependencies}{

\link{FrEDI} requires R (>= 4.2.0).

\link{FrEDI} depends on:
\itemize{
\item \link{tidyverse} (Easily Install and Load the 'Tidyverse'). The official documentation for \link{tidyverse} can be found \href{https://cran.r-project.org/web/packages/tidyverse/index.html}{here}. \link{tidyverse} can be installed using \code{install.packages("tidyverse")}, or see \href{https://tidyverse.tidyverse.org/}{link} for more information.
\item \link{ggpubr} ('ggplot2' Based Publication Ready Plots). The official documentation for \link{ggpubr} can be found \href{https://cran.r-project.org/web/packages/ggpubr/index.html}{here}. \link{ggpubr} can be installed using \code{install.packages("ggpubr")}, or see \href{https://rpkgs.datanovia.com/ggpubr/}{link} for more information.
\item \link{openxlsx} (Read, Write and Edit \code{xlsx} Files). The official documentation for \link{openxlsx} can be found \href{https://cran.r-project.org/web/packages/openxlsx/index.html}{here}. \link{openxlsx} can be installed using \code{install.packages("openxlsx")}.
}
}

\section{License}{

This repository is released under the MIT License.
}

\section{EPA Disclaimer}{

The United States Environmental Protection Agency (EPA) GitHub project code is provided on an "as is" basis and the user assumes responsibility for its use. EPA has relinquished control of the information and no longer has responsibility to protect the integrity, confidentiality, or availability of the information. Any reference to specific commercial products, processes, or services by service mark, trademark, manufacturer, or otherwise, does not constitute or imply their endorsement, recommendation or favoring by EPA. The EPA seal and logo shall not be used in any manner to imply endorsement of any commercial product or activity by EPA or the United States Government.

By submitting a pull request to the GitHub and/or by installing this package, you make an agreement with EPA that you will not submit a claim of compensation for services rendered to EPA or any other federal agency. Further, you agree not to charge the time you spend developing software code related to this project to any federal grant or cooperative agreement.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://usepa.github.io/FrEDI}
  \item \url{https://epa.gov/cira/FrEDI/}
  \item \url{https://github.com/USEPA/FrEDI/}
}

}
\author{
\strong{Maintainer}: Corinne Hartin \email{hartin.corinne@epa.gov} (\href{https://orcid.org/0000-0003-1834-6539}{ORCID}) (cahartin)

Authors:
\itemize{
  \item Erin McDuffie \email{mcduffie.erin.e@epa.gov} (\href{https://orcid.org/0000-0002-6845-6077}{ORCID}) (emcduffie) [contributor]
  \item Karen Noiva (\href{https://orcid.org/0000-0002-4412-5302}{ORCID}) (knoiva-indecon) [contributor]
}

Other contributors:
\itemize{
  \item Tony Gardella (\href{https://orcid.org/0000-0003-4380-3412}{ORCID}) (tonygard-indecon) [contributor]
  \item Will Maddock (\href{https://orcid.org/0000-0002-6107-0315}{ORCID}) (maddockw) [contributor]
  \item Hugh Easton (\href{https://orcid.org/0009-0008-0104-6891}{ORCID}) (heaston-indecon) [contributor]
}

}
\keyword{internal}
