% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrEDI.R
\docType{package}
\name{FrEDI}
\alias{FrEDI}
\title{README
FrEDI: The Framework for Evaluating Damages and Impacts}
\description{
\link{FrEDI} is an R package being developed by the U.S. Environmental Protection Agency (EPA). The functions and data provided by this package can be used to estimate climate change impacts for the contiguous United States (CONUS) using the Framework for Evaluating Damages and Impacts (FrEDI), developed as part of EPA's \href{https://epa.gov/cira/}{Climate Change Impacts and Risk Analysis} (CIRA) project. \link{FrEDI} contains R code that implement FrEDI and allow users to project impacts from climate change and sea level rise for a selected set of sectors.
}
\details{
For more information on the Framework and the CIRA project, visit https://epa.gov/cira/, especially https://epa.gov/cira/FrEDI.
}
\section{Overview of Functions}{

The function \code{\link[=run_fredi]{run_fredi()}} provided in this package is the primary function implementing FrEDI The main inputs to \code{\link[=run_fredi]{run_fredi()}} are climate scenarios (temperature in degrees Celsius, global mean sea level rise in centimeters) and socioeconomic scenarios (gross domestic product, regional population). \code{\link[=run_fredi]{run_fredi()}} projects climate impacts for these scenarios using the temperature binning framework. Users can also specify the levels at which to aggregate the climate impacts.

\link{FrEDI} also contains functions to assist in the pre-processing of input scenarios and the post-processing of outputs. For instance, the pre-processing function \code{\link[=import_inputs]{import_inputs()}} helps users to import scenarios from user-specified comma-separated value (CSV) files, while the post-processing function \code{\link[=get_plots]{get_plots()}} generates a couple of different types of images (creates heatmaps and ribbon plots of climate impacts over time) from the outputs of \code{\link[=run_fredi]{run_fredi()}}.

Additional helper functions provided by \link{FrEDI} include \code{\link[=convertTemps]{convertTemps()}}, \code{\link[=temps2slr]{temps2slr()}}, \code{\link[=get_sectorInfo]{get_sectorInfo()}}, and \code{\link[=aggregate_impacts]{aggregate_impacts()}}.

\code{\link[=convertTemps]{convertTemps()}} helps users to convert between global mean temperature and temperatures for the contiguous United States (CONUS) (both in degrees Celsius). \code{\link[=temps2slr]{temps2slr()}} helps users to estimate global mean sea level rise (GMSL, in centimeters) from global mean temperature in degrees Celsius.

\code{\link[=get_sectorInfo]{get_sectorInfo()}} allows users to access a list of sectors within FrEDI and related sector information.

The post-processing helper function \code{\link[=aggregate_impacts]{aggregate_impacts()}} helps users to aggregate and/or summarize the outputs of temperature binning (calculates national totals, model averages, impact totals, and interpolations between multiple impact years (for applicable sectors). Users have the option to run \code{\link[=aggregate_impacts]{aggregate_impacts()}} on the outputs of \code{\link[=run_fredi]{run_fredi()}}. However, users can alternatively specify aggregation levels as arguments to the main function, \code{\link[=run_fredi]{run_fredi()}}, which will run \code{\link[=aggregate_impacts]{aggregate_impacts()}} for specified levels before returning outputs (by default, \code{\link[=run_fredi]{run_fredi()}} returns results aggregated to all aggregation levels).

@section Overview of Package Contents:
\link{FrEDI} consists of files in the following directories:
\itemize{
\item \strong{R}. Contains function definitions (files ending in \code{".R"}) and configuration files (ending in \code{".rda"}).
\item \strong{data}. Contains R Data files ending in \code{".rdb"}, \code{".rds"}, and \code{".rdx"}, containing data included with the package.
\item \strong{help} and \strong{html}. Contain documentation for functions available to the user, including function descriptions, lists of arguments and outputs, and examples. See \code{"html/00Index.html"} or the individual R help functions for more information about individual functions.
\item \strong{Meta}. Contains RDS files (ending in \code{".rds"} with information about the package contents.
\item \strong{extdata/scenarios}. Contains three CSV files for users to test the function for importing data. (\code{"pop_scenario.csv"}, \code{"slr_from_GCAM.csv"}, and \code{"GCAM_scenario.csv"} respectively contain examples of a population scenario, sea level rise scenario, and temperature scenario. For more information, refer to documentation for the function \code{"import_inputs()"}.
}

The FrEDI R package contains a dataset with default results \code{defaultResults}, which contains annual impacts produced by \code{\link[=run_fredi]{run_fredi()}} for the default scenarios (i.e., default temperature, GDP and regional population trajectories), and can be loaded into the R workspace (\verb{load(defaultResults))}.

Typical use will involve \code{library(FrEDI)} or \code{require(FrEDI)}.
}

\section{Status}{

All code in this repository is being provided in a "draft" state and has not been reviewed or cleared by US EPA. This status will be updated as models are reviewed.
}

\section{Dependencies}{

\link{FrEDI} requires R (>= 3.5.0).

Installing \link{FrEDI} requires \link{devtools} (Tools to Make Developing R Packages Easier). More information on \link{devtools} can be found \href{https://cran.r-project.org/web/packages/devtools/index.html}{here.} \link{devtools} can be installed using \code{install.packages("devtools")}, or see \href{https://www.r-project.org/nosvn/pandoc/devtools.html}{link} for more information.

\link{FrEDI} depends on:
\itemize{
\item \link{tidyr} (Tidy Messy Data). The official documentation for \link{tidyr} can be found \href{https://cran.r-project.org/web/packages/tidyr/index.html}{here}. \link{tidyr} can be installed using \code{install.packages("tidyr")}, or see \href{https://tidyr.tidyverse.org/}{link} for more information.
\item \link{dplyr} (A Grammar of Data Manipulation). The official documentation for \link{dplyr} can be found \href{https://cran.r-project.org/web/packages/dplyr/index.html}{here}. dplyr can be installed using \code{install.packages("dplyr")}, or see \href{https://dplyr.tidyverse.org/}{link} for more information.
\item \link{ggplot2} (Create Elegant Data Visualisations Using the Grammar of Graphics). The official documentation for \link{ggplot2} can be found \href{https://cran.r-project.org/web/packages/ggplot2/index.html}{here}. \link{ggplot2} can be installed using \code{install.packages("ggplot2")}, or see \href{https://ggplot2.tidyverse.org/}{link} for more information.
}
}

\section{License}{

This repository is released under the MIT License.
}

\section{EPA Disclaimer}{

The United States Environmental Protection Agency (EPA) GitHub project code is provided on an "as is" basis and the user assumes responsibility for its use. EPA has relinquished control of the information and no longer has responsibility to protect the integrity, confidentiality, or availability of the information. Any reference to specific commercial products, processes, or services by service mark, trademark, manufacturer, or otherwise, does not constitute or imply their endorsement, recommendation or favoring by EPA. The EPA seal and logo shall not be used in any manner to imply endorsement of any commercial product or activity by EPA or the United States Government.

By submitting a pull request to the GitHub and/or by installing this package, you make an agreement with EPA that you will not submit a claim of compensation for services rendered to EPA or any other federal agency. Further, you agree not to charge the time you spend developing software code related to this project to any federal grant or cooperative agreement.
}

