% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots.R
\name{get_plots}
\alias{get_plots}
\title{Create plots of FrEDI outputs}
\usage{
get_plots(
  data = NULL,
  column = "annual_impacts",
  undiscounted = TRUE,
  plotTypes = "all",
  save = FALSE,
  directory = NULL,
  groupVars = c("sector", "variant")
)
}
\arguments{
\item{data}{Data frame of summarized outputs produced by \code{\link[=run_fredi]{run_fredi()}}. Do not change column names of the \code{\link[=run_fredi]{run_fredi()}} output before running \code{\link[=get_plots]{get_plots()}}.}

\item{column}{A character string indicating the name of the column in the data for which to create plots.}

\item{undiscounted}{A \code{TRUE/FALSE} value indicating whether the values in column represent undiscounted values or discounted values (i.e., present values). Defaults to \code{undiscounted=TRUE}.}

\item{plotTypes}{Character string or character vector indicating which types of plots to produce. Options are \code{c("heatmaps", "ribbon", "all")}. Set \code{plotTypes="all"} (default) to produce both types of plots.}

\item{save}{A \code{TRUE/FALSE} value indicating whether to save results. If a directory value is supplied (i.e., \code{!is.null(directory)}), defaults to \code{save=TRUE}. Otherwise, default is \code{save=FALSE}.}

\item{directory}{A character string indicating the location of a directory in which to save the plot objects. No default (i.e., \code{directory=NULL}).}

\item{groupVars}{A character vector indicating columns to use for grouping. Defaults to \code{groupVars=c("sector", "variant")}}
}
\value{
\code{\link[=get_plots]{get_plots()}} returns a nested list of plots:

\tabular{ll}{
\strong{List Index} \tab \strong{Description} \cr
\code{heatmaps} \tab List of heatmaps with list elements for all unique values for model types present in data (i.e., \code{"GCM"} for results calculated with temperature as the driver value and \code{"SLR"} for results calculated with GMSL as the driver value.\cr
\code{ribbon} \tab List of lists of annual results. List of annual plots contains a list of sectors. Each sector contains a list of nested regional plots (\code{regional}, with all regions except national results and \code{national}, for national results). \cr
}
}
\description{
`This function creates plots for the summarized FrEDI outputs. \code{\link[=get_plots]{get_plots()}} returns a list with heatmaps for model types present in the data (GCMs and SLR scenarios) and ribbon plots, by sector.
}
\details{
This function creates plots for the summarized FrEDI outputs. \code{\link[=get_plots]{get_plots()}} returns a list with heatmaps for model types present in the data (GCMs and SLR scenarios) and annual results for all sectors and variants. Results from FrEDI must be summed across impact types before using \code{\link[=get_plots]{get_plots()}} (use \code{\link[=run_fredi]{run_fredi()}} with the defaults or run \code{aggregate_impacts(aggLevels="impactType")} on the output from \code{\link[=run_fredi]{run_fredi()}}).

By default, \code{\link[=get_plots]{get_plots()}} plots results from the \code{"annual_impacts"} column (or users can specify a column name in the data with \code{column}).

The argument \code{undiscounted} is used by \code{\link[=get_plots]{get_plots()}} for plot labels and in file and directory names for saving results.

\code{\link[=get_plots]{get_plots()}} produces heatmaps (\code{plotTypes="heatmaps"}) for the outputs of FrEDI and/or plots the average value and range of impacts as a time series (\code{plotTypes="ribbon"}) for each sector-variant-region combination. Users can specify which plot types to produce by setting \code{plotTypes}. Set \code{plotTypes="all"} (default) to produce both heat maps and annual results or specify a single type (\code{plotTypes="heatmaps"} and \code{plotTypes="ribbon"}, respectively).

The heatmaps display the numeric values in the specified column (e.g., \code{"annual_impacts"}) as a grid of colored pixels. Each row in the grid corresponds to a sector-variant combination (e.g., "Coastal Properties, No Adaptation"), while columns in the grid correspond to years. In other words, the heatmaps display the relative intensity of the impacts of a sector and variant compared to others. The colors in the heatmap are a gradient ranging from dark blue (impacts with values below zero) to dark red (impacts with values above zero), with a midpoint at zero (missing values appear as grey pixels). The scale of the gradient is determined from the underlying data, with the darkest points corresponding to the minimum and maximum values.

If temperature-driven (GCM sectors) and SLR-driven (SLR sectors) sectors are both present in the data, \code{\link[=get_plots]{get_plots()}} will produce a separate heatmap for each. Each heatmap displays panels for each region (stacked vertically) and underlying models (organized horizontally).

Setting (\code{plotTypes="ribbon"}) plots the annual impacts as time series. For temperature-driven sectors, the model average is plotted as a line and the range of model values (minimum and maximum) are plotted as a ribbon plot. For the SLR-driven sectors, the interpolated impacts are plotted as a line. For sectors with multiple variants, impacts for individual variants are displayed in separate panels (organized horizontally).

If \code{save=TRUE} and the user supplies a path to a directory (i.e., \code{!is.null(directory)}), \code{\link[=get_plots]{get_plots()}} will try to save results to the specified directory. Separate directories are created within the specified directory for heatmaps and ribbon plots.
}
\examples{
### Create temperature binning scenario
df_tempExOut <- get_fredi(aggLevels="none", pv=TRUE, silent=TRUE)

### Aggregate temperature binning summary across multiple columns
agg_tempExOut <- df_tempExOut \%>\% aggregate_impacts(columns=c("annual_impacts", "discounted_impacts"))

### Create list of plots for aggregated results
agg_plotList <- agg_tempExOut \%>\% get_plots()

### Create list of heatmaps for regional values only:
reg_plotList <- agg_tempExOut \%>\% filter(region!="National Total") \%>\% get_plots(plotTypes="heatmaps")

### Create list of annual plots for national values only:
nation_plotList <- agg_tempExOut \%>\% filter(region=="National Total") \%>\% get_plots(plotTypes="annual")

}
\references{
Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at \url{https://epa.gov/cira/FrEDI/}.
}
