% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sectorInfo.R
\name{get_sectorInfo}
\alias{get_sectorInfo}
\title{This function provides information about the sectors in FrEDI.}
\usage{
get_sectorInfo(description = F, gcmOnly = F, slrOnly = F)
}
\arguments{
\item{description}{= FALSE. Logical value indicating whether to include information about each sector. Returns a dataframe if \code{description=TRUE} and returns a character vector of sector names if \code{description=FALSE} (default).}

\item{gcmOnly}{= FALSE. Logical value indicating whether to return only sectors with climate impacts modeled using global climate model (GCM) results.}

\item{slrOnly}{= FALSE. Logical value indicating whether to return only sectors with climate impacts modeled using sea level rise (SLR) scenarios.}
}
\value{
\itemize{
\item If \code{description=FALSE} (default), outputs a character vector containing the names of sectors available for FrEDI.
\item If \code{description=TRUE}, `, outputs a dataframe containing the names of sectors available for FrEDI in one column, with information about the sector model type, adaptations, impact years, and impact types in the remaining columns.
}
}
\description{
This helper function returns a character vector of sector names (default) \emph{or} a dataframe of sectors with related information (e.g., associated adaptations, impact types, etc.), which the user can supply to the \code{\link[=run_fredi]{run_fredi()}} \code{sectorList} argument.
}
\details{
If \code{description=FALSE} (default), this helper function returns a character vector of sector names, which the user can supply to the \code{sectorList} argument to \code{\link[=run_fredi]{run_fredi()}}. If \code{description=TRUE}, \code{get_sectorInfo()} returns dataframe of sectors with related information returns a dataframe containing the sectors available for FrEDI along with additional information. Sector names are in the first column, with additional columns for the associated model type ("GCM" or "SLR"), adaptations, impact years, and impact types in the remaining columns. Adaptations, impact years, and impact types vary by sector.

Users can specify whether to return only GCM sectors \emph{or} SLR sectors by setting \code{gcmOnly=TRUE} or \code{slrOnly=TRUE}, respectively. \code{\link[=get_sectorInfo]{get_sectorInfo()}} will return the sectors in the form specified by \code{description} (see above).
}
\examples{

### Return a character vector with the names of all of the sectors in FrEDI:
get_sectorInfo()

### Return a dataframe of all of the sectors in FrEDI (sector names and additional information)
get_sectorInfo(description=T, gcmOnly=T)

### Return a character vector with only the names of the temperature-driven sectors:
get_sectorInfo(gcmOnly=T)

### Run FrEDI for only the temperature-driven sectors and view results:
df_x <- run_fredi(sectorList=get_sectorInfo(gcmOnly=T))


}
\references{
Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at \url{https://epa.gov/cira/FrEDI/}.
}
