% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sv.R
\name{get_sv_sectorInfo}
\alias{get_sv_sectorInfo}
\title{Retrieve a character vector of sectors available in the FrEDI Social Vulnerability (SV) module (\link{run_fredi_sv}) or a data frame with SV sectors and additional information.}
\usage{
get_sv_sectorInfo(description = F, gcmOnly = F, slrOnly = F)
}
\arguments{
\item{description}{Logical value indicating whether to include information about each sector. Returns a data frame if \code{description=TRUE} and returns a character vector of sector names if \code{description=FALSE} (default).}

\item{gcmOnly}{Logical value indicating whether to return only sectors with climate impacts modeled using global climate model (GCM) results.}

\item{slrOnly}{Logical value indicating whether to return only sectors with climate impacts modeled using sea level rise (SLR) scenarios.}
}
\value{
\code{\link[=get_sv_sectorInfo]{get_sv_sectorInfo()}} will return SV sectors in the form specified by \code{description}:
\itemize{
\item If \code{description=FALSE} (default), outputs a character vector containing the names of sectors available for the FrEDI SV Module.
\item If \code{description=TRUE}, outputs a data frame containing the names of sectors available for the FrEDI SV Module in one column, with information about the sector model type, model type ("GCM" or "SLR") and associated driver unit ("degrees Celsius" or "cm", respectively ), impact units (e.g., mortality, etc.), and available variants in the remaining columns.
}
}
\description{
\code{get_sv_sectorInfo} returns a character vector with the names of sectors in the FrEDI Social Vulnerability (SV) module (default) \strong{or} a data frame of SV sectors with additional information (e.g., associated variants, model type, etc.).
}
\details{
\code{get_sv_sectorInfo} returns a character vector with the names of sectors in the FrEDI SV module (\code{description=FALSE}, default) \strong{or} a data frame with SV sectors and additional information (\code{description=TRUE}).
\itemize{
\item If \code{description=FALSE} (default), \code{get_sv_sectorInfo} returns a character vector of names of sectors that can be passed to the \code{\link[=run_sv_fredi]{run_sv_fredi()}} \code{sector} argument. Specify whether to return only GCM sectors by running \code{get_sv_sectorInfo(gcmOnly=TRUE)} or SLR sectors by running \code{get_sv_sectorInfo(slrOnly=TRUE)}.
\item If \code{description=TRUE}, \code{get_sv_sectorInfo} returns a data frame of sectors with related information, such as whether a particular sector is driven primarily by temperature (\code{modelType="GCM"}) or sea level rise (\code{modelType="SLR"}), associated driver units (\code{"degrees Celsius"} for temperature-driven sectors, \code{"cm"} for SLR-driven sectors), impact units (e.g., mortality, etc.), and sector variants. Users can use \link{get_sv_sectorInfo} to determine which sectors can be passed to the \code{\link[=run_sv_fredi]{run_sv_fredi()}} \code{sector} argument and/or to determine whether a particular sector is driven primarily by temperature (\code{modelType="GCM"}) or sea level rise (\code{modelType="SLR"}).
}

Users can specify whether to return only GCM sectors by setting \code{gcmOnly=TRUE} \strong{or} SLR sectors by setting \code{slrOnly=TRUE}.
}
\examples{

### Return a character vector with the names of all of the sectors in the FrEDI SV Module:
get_sv_sectorInfo()

### Return a data frame of all of the sectors in the FrEDI SV Module (sector names and additional information)
get_sv_sectorInfo(description=T)

### Return a character vector with only the names of the temperature-driven sectors:
get_sv_sectorInfo(gcmOnly=T)

### Return a character vector with only the names of the temperature-driven sectors:
get_sv_sectorInfo(slrOnly=T)

}
\references{
Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at \url{https://epa.gov/cira/FrEDI/}.
}
