% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_inputs.R
\name{import_inputs}
\alias{import_inputs}
\title{Import custom scenarios for temperature, global mean sea level rise (GMSL), population, and GDP from user-specified file names}
\usage{
import_inputs(
  tempfile = NULL,
  slrfile = NULL,
  popfile = NULL,
  gdpfile = NULL,
  temptype = "conus",
  popform = "wide",
  byState = FALSE
)
}
\arguments{
\item{tempfile}{A character string indicating the location of a CSV file containing a custom temperature scenario (first column contains years; second column contains temperatures, in degrees Celsius, above the 1995 baseline year). The temperature scenario must start in 2000 or earlier and end at or after the maximum model run year (e.g., as specified by the \code{maxYear} argument to \code{\link[=run_fredi]{run_fredi()}}).}

\item{slrfile}{A character string indicating the location of a CSV file containing a custom sea level rise scenario (first column contains years; second column contains values for global mean sea level rise (GMSL), in centimeters, above the 2000 baseline). The SLR scenario must start in 2000 or earlier and end at or after the maximum model run year (e.g., as specified by the \code{maxYear} argument to \code{\link[=run_fredi]{run_fredi()}}).}

\item{popfile}{A character string indicating the location of a CSV file containing a custom population scenario for states and NCA regions. The first column contains years in the interval 2010 to 2300. The second column should contain the NCA Region label associated with the state. The third column should contain state names. The fourth column should contain the state postal code abbreviation. The fifth column should contain the population values. The population scenario must start in 2010 or earlier and end at or after the maximum model run year (e.g., as specified by the \code{maxYear} argument to \code{\link[=run_fredi]{run_fredi()}}).}

\item{gdpfile}{A character string indicating the location of a CSV file containing a custom scenario for gross domestic product (GDP) (first column contains years; second column contains values for GDP, in total 2015$). The GDP scenario must start in 2010 or earlier and end at or after the maximum model run year (e.g., as specified by the \code{maxYear} argument to \code{\link[=run_fredi]{run_fredi()}}).}

\item{temptype}{A character string indicating whether the temperature values in the temperature input file (specified by \code{tempfile} represent global temperature change (\code{temptype="global"}) or temperature change for the contiguous U.S. (\code{temptype="conus"}) in degrees Celsius. By default, the model assumes temperatures are CONUS temperatures (i.e., \code{temptype="conus"}).}
}
\value{
\code{\link[=import_inputs]{import_inputs()}} returns a list of named elements containing data frames with custom scenarios for temperature, GMSL, GDP, and regional population, respectively:

\tabular{ll}{
\strong{List Index} \tab \strong{Description} \cr
\code{tempInput} \tab Data frame containing a custom temperature scenario imported from the CSV file specified by \code{tempfile}, with missing values removed. \code{tempInput} has two columns with names \code{"year"} and \code{"temp_C"}, containing the year and CONUS temperatures in degrees Celsius, respectively. \cr
\code{slrInput} \tab Data frame containing a custom GMSL scenario imported from the CSV file specified by \code{slrfile}, with missing values removed. \code{slrInput} has two columns with names \code{"year"}, and \code{"slr_cm"}, containing the year and global mean sea level rise (GMSL) in centimeters, respectively. \cr
\code{gdpInput} \tab Data frame containing a custom GDP scenario imported from the CSV file specified by \code{gdpfile}, with missing values removed. \code{gdpInput} has two columns with names \code{"year"}, and \code{"gdp_usd"}, containing the year and the national GDP, respectively. \cr
\code{popInput} \tab Data frame containing a custom temperature scenario imported from the CSV file specified by \code{popfile}, with missing values removed. \code{popInput} has three columns with names \code{"year"}, \code{"region"}, \code{"state"}, \code{"postal"}, and \code{"reg_pop"}, containing the year, the NCA region name, and the state, the postal code abbreviation, and the state population, respectively.. \cr
}
}
\description{
This function enables users to import data on custom scenarios for use with \link{FREDI} (supplied as inputs to \code{\link[=run_fredi]{run_fredi()}}). Users specify path names to CSV files containing temperature, global mean sea level rise (GMSL), gross domestic product (GDP), and population scenarios. \code{\link[=import_inputs]{import_inputs()}} reads in and format any specified files as data frames and returns a list of data frames for imported scenarios.
}
\details{
This function enables users to import data on custom scenarios for use with temperature binning. Users specify path names to CSV files containing temperature, global mean sea level rise (GMSL), population, and gross domestic product (GDP) scenarios (\code{tempfile}, \code{slrfile}, \code{gdpfile}, and \code{popfile}, respectively). \code{\link[=import_inputs]{import_inputs()}} reads in and formats any specified files as data frames and returns a list of data frames for imported scenarios. Users can specify whether the temperature input is for the contiguous U.S. (CONUS) or global using \code{temptype}.
\itemize{
\item \strong{Temperature Inputs.} The input temperature scenario requires CONUS temperatures in degrees Celsius relative to 1995 (degrees of warming relative to the baseline year--i.e., the central year of the 1986-2005 baseline). CONUS temperature values must be greater than or equal to zero degrees Celsius.
\itemize{
\item Users can convert global temperatures to CONUS temperatures using \link{convertTemps}\code{(from="global")} (or by specifying \link{import_inputs}\code{(temptype="global")} when using \code{\link[=import_inputs]{import_inputs()}} to import a temperature scenario from a CSV file).
\item Temperature inputs must have at least one non-missing value in 2000 or earlier and at least one non-missing value in or after the final analysis year (as specified by the \code{\link[=run_fredi]{run_fredi()}} \code{maxYear} argument).
}
\item \strong{SLR Inputs.} The input SLR scenario requires values for changes in global mean sea level rise (GMSL) heights in centimeters (cm). GMSL heights must be greater than or equal to zero.
\itemize{
\item \code{slrInput} requires a data frame object with two columns containing the year and global mean sea level rise (GMSL) in centimeters, respectively.
\item SLR inputs must have at least one non-missing value in 2000 or earlier and at least one non-missing value in or after the final analysis year (as specified by the \code{\link[=run_fredi]{run_fredi()}} \code{maxYear} argument).
}
\item \strong{GDP Inputs.} The input scenario for gross domestic product (GDP) requires national GDP values in 2015$. GDP values must be greater than or equal to zero.
\itemize{
\item \code{gdpInput} requires a data frame object with five columns with names \code{"year"}, and \code{"gdp_usd"}, containing the year and the national GDP, respectively. GDP values must be greater than or equal to zero.
\item GDP inputs must have at least one non-missing value in 2010 or earlier and at least one non-missing value in or after the final analysis year (as specified by the \code{\link[=run_fredi]{run_fredi()}} \code{maxYear} argument).
}
\item \strong{Population Inputs.} The input population scenario requires state-level population values. Population values must be greater than or equal to zero.
\itemize{
\item \code{popInput} requires a data frame object with five columns with names \code{"year"}, \code{"region"}, \code{"state"}, \code{"postal"}, and \code{"reg_pop"}, containing the year, the NCA region name, and the state, the postal code abbreviation, and the state population, respectively.
\item Population inputs must have at least one non-missing value in 2010 or earlier and at least one non-missing value in or after the final analysis year (as specified by the \code{\link[=run_fredi]{run_fredi()}} \code{maxYear} argument).
}
}

\code{\link[=import_inputs]{import_inputs()}} outputs a list of data frames that can be passed to the main FREDI function \code{\link[=run_fredi]{run_fredi()}} using the \code{inputList} argument. For example, specify \code{run_fredi(inputsList=x)} to generate impacts for a custom scenario \code{x} (where \code{x} is a list of data frames such as that output from \code{\link[=import_inputs]{import_inputs()}}) (see \code{\link[=run_fredi]{run_fredi()}}).

All inputs to \code{\link[=import_inputs]{import_inputs()}} are optional. If the user does not specify a file path for \code{tempfile}, \code{slrfile}, \code{gdpfile}, or \code{popfile} (or if there is an error reading in inputs from those file paths), \code{\link[=import_inputs]{import_inputs()}} outputs a list with a \code{NULL} value for the associated list element. \code{\link[=run_fredi]{run_fredi()}} defaults back to the default scenarios for any elements of the inputs list that are \code{NULL} or missing. In other words, running \code{run_fredi(inputsList=list())} returns the same outputs as running \code{\link[=run_fredi]{run_fredi()}} (see \code{\link[=run_fredi]{run_fredi()}}).
}
\examples{
### Path to example scenarios
scenariosPath <- system.file(package="FrEDI") |> file.path("extdata","scenarios")

### View example scenario names
scenariosPath |> list.files()

### Temperature Scenario File Name
tempInputFile <- scenariosPath |> file.path("GCAM_scenario.csv")

### SLR Scenario File Name
slrInputFile  <- scenariosPath |> file.path("slr_from_gcam.csv")

### Population Scenario File Name
popInputFile  <- scenariosPath |> file.path("pop_scenario.csv")

### Import inputs
example_inputsList <- import_inputs(
  tempfile = tempInputFile,
  slrfile  = slrInputFile,
  popfile  = popInputFile,
  temptype = "global",
  popform  = "wide"
)

### Use imports with FREDI:
df_x <- run_fredi(inputsList=example_inputsList)

}
\references{
Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at \url{https://epa.gov/cira/FrEDI/}.
}
