% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popDefault}
\alias{popDefault}
\title{U.S. state population scenario, which can be passed as an input to \code{\link[=run_fredi]{run_fredi()}}, \code{\link[=run_fredi_sv]{run_fredi_sv()}}, and/or \code{\link[=FrEDI:run_fredi_ghg]{FrEDI:run_fredi_ghg()}}.}
\format{
A data frame with 4,459 rows and 5 columns:
\describe{
\item{region}{Region of the U.S.: one of \code{"Midwest"}, \code{"Northeast"}, \code{"Northern Plains"}, \code{"Northwest"}, \code{"Southeast"}, \code{"Southern Plains"}, or \code{"Southwest"} for CONUS states, or \code{"Alaska"} and \code{"Hawaii"} for Alaska and Hawaii, respectively.}
\item{state }{U.S. states or the District of Columbia}
\item{postal}{Two-letter postal code abbreviation associated with the state (e.g., \code{"AL"} for Alabama)}
\item{year  }{Year}
\item{pop   }{State population for associated region and year}
}
}
\source{
\url{https://epa.gov/cira/FrEDI/}
}
\usage{
popDefault
}
\description{
\code{popDefault} is a data frame object that contains population projections for 50 U.S. states and the District of Columbia for the period from 2010 to 2100. Values for the 48 states and the District of Columbia comprising the contiguous U.S. (CONUS) are from the (\strong{ICLUSv2}) model (Bierwagen et al, 2010; EPA 2017) under the Median variant projection of United Nations (United Nations, 2015). Values for Alaska and Hawaii are from the U.S. Census Bureau.
}
\details{
To use \code{popDefault} as an input to FrEDI, the FrEDI SV module, and/or the FrEDI Methane module, pass the data frame via the named element \code{pop} in a list passed to the \code{inputsList} argument in function calls to \code{\link[=run_fredi]{run_fredi()}}, \code{\link[=run_fredi_sv]{run_fredi_sv()}}, and/or \code{\link[=FrEDI:run_fredi_ghg]{FrEDI:run_fredi_ghg()}}, respectively, e.g.: \code{run_fredi(inputsList=list(pop=popDefault))}. \code{popDefault} can also be combined with other provided scenarios (\code{gcamScenarios}, \code{gdpDefault}) in function calls (e.g., \code{run_fredi(inputsList=list(gdp=gdpDefault, pop=popDefault))}) or user-provided data frames. For more information, see documentation for \code{\link[=run_fredi]{run_fredi()}}, \code{\link[=FrEDI:run_fredi_sv]{FrEDI:run_fredi_sv()}}, \code{\link[=FrEDI:run_fredi_ghg]{FrEDI:run_fredi_ghg()}}, and \code{\link[=import_inputs]{import_inputs()}}.

Bierwagen, B., D. M. Theobald, C. R. Pyke, A. Choate, P. Groth, J. V. Thomas, and P. Morefield. 2010. \url{https://doi.org/10.1073/pnas.1002096107}.

EPA. 2017. \url{https://doi.org/10.13140/RG.2.2.14466.79045}

United Nations. 2015. World population prospects: The 2015 revision. New York: United Nations, Department of Economic and Social Affairs, Population Division.
}
\keyword{datasets}
