% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fredi_sv.R
\name{run_fredi_sv}
\alias{run_fredi_sv}
\title{FrEDI Social Vulnerability (SV) Module: calculates climate change impacts on socially vulnerable populations throughout the 21st century for available sectors}
\usage{
run_fredi_sv(
  inputsList = list(driverInput = NULL, popInput = NULL),
  sector = NULL,
  silent = TRUE,
  driverInput = NULL,
  popInput = NULL,
  save = FALSE,
  outpath = getwd(),
  overwrite = FALSE,
  addDate = FALSE
)
}
\arguments{
\item{sector=NULL}{A character string indicating for which sector to run the FrEDI SV module.}

\item{driverInput=NULL}{A dataframe of up to four custom scenarios for drivers (temperature or global mean sea level rise). \code{driverInput} requires a dataframe with columns of \code{"year"} and \code{"scenario"}. The dataframe must also include at least one of the following columns: \code{"temp_C"} or \code{"slr_cm"} (depending on whether the sector impacts are driven primarily by temperature or sea level rise, respectively...use \code{\link[=get_sv_sectorInfo]{get_sv_sectorInfo()}} to get information on driver types). If users include all four columns (\verb{c(``"year",} \verb{"scenario",} \verb{"temp_C",} \verb{"slr_cm"``)}) then \code{\link[=run_fredi_sv]{run_fredi_sv()}} will determine which driver column (\code{"temp_C"} or \code{"slr_cm"}) to use based on the specified sector. Driver inputs (temperature and sea level rise) should start in the year 2000 or earlier. Temperature inputs must be temperature change in degrees Celsius for the CONUS region (if starting from global temperature change, use \code{\link[=convertTemps]{convertTemps()}} to convert global temperatures to CONUS temperatures before passing to \code{driverInput}). Sea level rise inputs must be change in sea level in centimeters. All scenarios must include at least two non-missing values. If any required columns are missing, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will use the default temperature scenario from \code{\link[=run_fredi]{run_fredi()}}. If the dataframe passed to \code{driverInput} has more than four unique scenarios, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will only run the first four scenarios.}

\item{popInput=NULL}{A dataframe containing a custom scenario for regional population, with columns \verb{c(``"year",} \verb{"region",} \verb{"reg_pop"``)} (containing the year, region, and regional population, respectively). The dataframe passed to \code{popInput} can be imported using \code{\link[=import_inputs]{import_inputs()}} (for more information, see \code{\link[=import_inputs]{import_inputs()}}). Note that, in contrast to the dataframe passed to \code{driverInput}, the \code{popInput} dataframe must be a single scenario (i.e., \code{\link[=run_fredi_sv]{run_fredi_sv()}} uses the same population scenario for all driver scenarios in \code{driverInput}). Region names in the \code{"region"} column must match those in \verb{c(``"Midwest",} \verb{"Northeast",} \verb{"Northwest",} \verb{"Northern Plains",} \verb{"Southeast",} \verb{"Southwest",} \verb{"Southern Plains"``)} or \verb{c(``"Midwest",} \verb{"Northeast",} \verb{"Northwest",} \verb{"Northern.Plains",} \verb{"Southeast",} \verb{"Southwest",} \verb{"Southern.Plains"``)}.}

\item{silent=TRUE}{A \code{TRUE/FALSE} value indicating the level of messaging desired by the user (default=\code{TRUE}).}

\item{save=FALSE}{A \code{TRUE/FALSE} value indicating whether to save the results to an Excel file (default=\code{FALSE}).}

\item{outpath=getwd()}{A character string indicating a file directory to save the Excel file (created in \code{save=TRUE}). By default, if \code{save=TRUE}, the Excel file will be saved to the working directory (with a file name determined by the sector).}

\item{overwrite=FALSE}{A \code{TRUE/FALSE} value indicating whether to overwrite an existing Excel file if \code{save=TRUE} (default=\code{FALSE}). By default, if \code{save=TRUE}, \code{overwrite=FALSE}. If \code{overwrite=FALSE}, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will not write over an existing file. If \code{overwrite=FALSE}, and the file already exists, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will provide the user with the option to write over the existing file: in this case, when \code{\link[=run_fredi_sv]{run_fredi_sv()}} is ready to save the file, it will message the user and prompt them if they would like to write over the existing file.}

\item{addDate=FALSE}{A \code{TRUE/FALSE} value indicating whether to add the date to the name of the output Excel file if \code{save=TRUE} (\code{default=FALSE}). By default, if \code{save=TRUE}, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will not add the date to the Excel file name (\code{addDate=FALSE}). If \code{addDate=TRUE} (and \code{save=TRUE}), \code{\link[=run_fredi_sv]{run_fredi_sv()}} will append the system date to the beginning of the name of the outputs Excel file using the format \code{"\%Y\%m\%d"} (see \code{\link[base:format]{base::format()}} and \code{\link[base:Sys.time]{base::Sys.Date()}} for additional information).}
}
\value{
The output of \code{\link[=run_fredi_sv]{run_fredi_sv()}} is an R data frame object containing annual average impacts, by year (2010-2090), for each sector, adaptation, model (GCM or SLR scenario), and region.
}
\description{
This function allows users to project annual average climate change impacts throughout the 21st century (2010-2090) for socially vulnerable populations for available sectors (see \code{\link[=get_sv_sectorInfo]{get_sv_sectorInfo()}}). Users must pass a dataframe of custom driver scenarios to \code{\link[=run_fredi_sv]{run_fredi_sv()}} using the \code{driverInput} argument (required) and have the option to pass a population scenario to \code{\link[=run_fredi_sv]{run_fredi_sv()}} via the \code{popInput} argument (optional). The output is an R data frame object containing annual average impacts at five-year increments between 2010 and 2090. Users have the option to write results to an Excel file by setting \code{save=TRUE}. Additional arguments provide more control over how the outputs are saved if \code{save=TRUE}.
}
\details{
This function allows users to project annual average climate change impacts throughout the 21st century (2010-2090) for socially vulnerable populations for available sectors (see \code{\link[=get_sv_sectorInfo]{get_sv_sectorInfo()}}). \code{\link[=run_fredi_sv]{run_fredi_sv()}} is the main function for the FrEDI Social Vulnerability (SV) Module in the \link{FrEDI} R package, described elsewhere (See \url{https://epa.gov/cira/FrEDI} for more information). Users have the option to pass a dataframe of custom driver scenarios to \code{\link[=run_fredi_sv]{run_fredi_sv()}} using the \code{driverInput} argument (required) and have the option to pass a population scenario to \code{\link[=run_fredi_sv]{run_fredi_sv()}} via the \code{popInput} argument (optional). The output is an R data frame object containing annual average impacts at five-year increments between 2010 and 2090. Users have the option to write results to an Excel file by setting \code{save=TRUE}. Additional arguments provide more control over how the outputs are saved if \code{save=TRUE}.

\code{driverInput} requires a dataframe of up to four custom scenarios for drivers (temperature or global mean sea level rise). \code{driverInput} requires a dataframe with columns of \code{"year"} and \code{"scenario"}. The dataframe must also include at least one of the following columns: \code{"temp_C"} or \code{"slr_cm"} (depending on whether the sector impacts are driven primarily by temperature or sea level rise, respectively). If users include all four columns (\verb{c(``"year",``"scenario",``"temp_C",``"slr_cm"``)}) then \code{\link[=run_fredi_sv]{run_fredi_sv()}} will determine which driver column (\code{"temp_C"} or \code{"slr_cm"}) to use based on the specified sector. Driver inputs (temperature and sea level rise) should start in the year 2000 or earlier. Temperature inputs must be temperature change in degrees Celsius for the CONUS region (if starting from global temperature change, use \code{\link[=convertTemps]{convertTemps()}} to convert global temperatures to CONUS temperatures before passing to \code{driverInput}). Sea level rise inputs must be change in sea level in centimeters. All scenarios must include at least two non-missing values. If any required columns are missing, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will use the default temperature scenario from \code{\link[=run_fredi]{run_fredi()}}. If the dataframe passed to \code{driverInput} has more than four unique scenarios, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will only run the first four scenarios.

\code{popInput} is an optional input that takes a dataframe containing a custom scenario for regional population, with columns \verb{c(``"year",} \verb{"region",} \verb{"reg_pop"``)} (containing the year, region, and regional population, respectively). If \code{popInput=NULL} (default), \code{\link[=run_fredi_sv]{run_fredi_sv()}} will use the a default regional population scenario from ICLUS (see \url{https://epa.gov/cira/FrEDI} for more information about the ICLUS scenario). The dataframe passed to \code{popInput} can be imported using \code{\link[=import_inputs]{import_inputs()}} (for more information, see \code{\link[=import_inputs]{import_inputs()}}). Note that, in contrast to the dataframe passed to \code{driverInput}, the \code{popInput} dataframe must be a single scenario (i.e., \code{\link[=run_fredi_sv]{run_fredi_sv()}} uses the same population scenario for all driver scenarios in \code{driverInput}). Region names in the \code{"region"} column must match those in \verb{c(``"Midwest",} \verb{"Northeast",} \verb{"Northwest",} \verb{"Northern Plains",} \verb{"Southeast",} \verb{"Southwest",} \verb{"Southern Plains"``)} or \verb{c(``"Midwest",} \verb{"Northeast",} \verb{"Northwest",} \verb{"Northern.Plains",} \verb{"Southeast",} \verb{"Southwest",} \verb{"Southern.Plains"``)}.
}
\examples{
### Run function with defaults (same as `defaultResults` dataset)
sv_defaults <- run_fredi_sv()

### Path to example scenarios
scenariosPath <- system.file(package="FrEDI") \%>\% file.path("extdata","scenarios")
### View example scenario names
scenariosPath \%>\% list.files
### Temperature Scenario File Name
tempInputFile <- scenariosPath \%>\% file.path("GCAM_scenario.csv")
### SLR Scenario File Name
slrInputFile  <- scenariosPath \%>\% file.path("slr_from_GCAM.csv")
### Population Scenario File Name
popInputFile  <- scenariosPath \%>\% file.path("pop_scenario.csv")
### Import inputs
example_inputsList <- import_inputs(
  tempfile = tempInputFile,
  slrfile  = slrInputFile,
  popfile  = popInputFile
)

### Run custom temperature scenario and output impacts without aggregation and with present values (default base year and discount rate)
df_tempExOut <- run_fredi(inputsList= tempBin_inputs, aggLevels="none", pv=TRUE, silent=TRUE)

}
\references{
Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at \url{https://epa.gov/cira/FrEDI/}.
}
