% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temps2slr.R
\name{temps2slr}
\alias{temps2slr}
\title{Convert global temperature change in degrees Celsius to global mean sea level rise (GMSL) in centimeters}
\usage{
temps2slr(temps, years)
}
\arguments{
\item{temps}{A numeric vector of global temperatures in degrees Celsius. The temperature series and corresponding years must begin in 2000 or earlier.}

\item{years}{A numeric vector of years (common era) corresponding to the temperatures provided to \code{temps}. The temperature series and corresponding years must begin in 2000 or earlier.}
}
\value{
Outputs a dataframe with two columns: \code{year}, which has the years from the \code{years} input that fall within the range from 2000 through 2090 and \code{slr_cm} which has the GMSL in centimeters.
}
\description{
This pre-processing helper function converts a vector of global temperatures to global mean sea level rise (GMSL). This function adapts the method described by Kopp et al. (2016) (see references, below).
}
\details{
This function converts a temperature scenario (global temperatures in degrees Celsius) into an SLR scenario to use for estimate impacts of global mean sea level rise (GMSL) on affected sectors. \code{\link[=temps2slr]{temps2slr()}} implements the method described by Kopp et al., 2016, "Temperature-driven global sea-level variability in the Common Era" (see references, below).

Kopp et al.'s "semiempirical" sea-level model relates the rate of global sea level (GSL) rise \eqn{\frac{dh}{dt}}{dh/dt} to global mean temperature \eqn{T(t)}, an equilibrium temperature \eqn{T_0(t)}{Te(t)}, and a small residual trend arising from the long-term response to earlier climate change \eqn{\phi(t)}{\phi(t)}:

\deqn{dh/dt = \alpha \cdot (T(t) - T_0(t)) + \phi(t)}{dh/dt = \alpha \cdot (T(t) - Te(t)) + \phi(t)}

As explained by Kopp et al. (2016): "The first term describes the GSL response to climate change during the study period. The second term covers a small residual trend arising from the long-term response to earlier climate change (i.e., deglaciation), which is very slowly decaying over millennia and of the order 0.1 mm/year in 2000 CE. It thus has a negligible effect on the modeled GSL rise during the 20th and 21st centuries" (Kopp et al., 2016, p. E1439).

In the equation above, \eqn{T_0(t)}{Te(t)} and \eqn{\phi(t)} are functions of time, where:

\deqn{\frac{dT_0(t)}{dt} = \frac{T(t) - Te(t)}{\tau_1}}{dTe(t)/dt = (T(t) - Te(t))/k1}

And:

\deqn{\frac{d\phi(t)}{dt} = \frac{\phi}{\tau_2}}{d\phi(t)/dt = p(t)/k2}

And \eqn{\tau_1}{k1} and \eqn{\tau_2}{k2} are both constants.
}
\examples{

### Path to example scenarios
scenariosPath <- system.file(package="FrEDI") \%>\% file.path("extdata","scenarios")
### View example scenario names

scenariosPath \%>\% list.files

### Temperature Scenario File Name
tempInputFile <- scenariosPath \%>\% file.path("GCAM_scenario.csv")

### Import example temperature scenario
example_inputsList <- import_inputs(tempfile = tempInputFile)

### Extract the example temperature scenario dataframe from the list
### Example has global temperatures in degrees Celsius
x_tempInput <- example_inputsList$tempInput

### Calculate global mean sea level rise in cm from global temperatures
x_slr <- temps2slr(temps=x_tempInput$temp_C, years=x_tempInput$year)


}
\references{
Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at \url{https://epa.gov/cira/FrEDI/}.

Kopp, Robert E., et al. (2016). Temperature-driven global sea-level variability in the Common Era. PNAS: E1434-E1441. Available at https://www.pnas.org/content/113/11/E1434
}
