% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_impacts.R
\name{aggregate_impacts}
\alias{aggregate_impacts}
\title{Summarize and aggregate impacts from FrEDI (calculate national totals, average across models, sum impact types, and interpolate between impact year estimates)}
\usage{
aggregate_impacts(
  data,
  aggLevels = c("national", "modelaverage", "impactyear", "impacttype"),
  columns = c("annual_impacts"),
  groupByCols = c("sector", "variant", "impactYear", "impactType", "model_type",
    "model", "region"),
  mode = "all"
)
}
\arguments{
\item{data}{Data frame of results FrEDI (outputs from \code{\link[=run_fredi]{run_fredi()}})}

\item{aggLevels}{Levels of aggregation at which to summarize data: one or more of \verb{c("national"}, \code{"modelAverage"}, \code{"impactYear"}, \code{"impactType"}, \verb{"all")}. Defaults to all levels (i.e., \code{aggLevels="all"}).}

\item{columns}{Character vector of columns for which to aggregate results (defaults to \verb{columns=c("annual_impacts"})).}

\item{groupByCols}{Character vector indicating which columns to use for grouping. Defaults to \verb{groupByCols=}c("sector"\verb{, }"variant"\verb{, }"impactYear"\verb{, }"impactType"\verb{, }"model_type"\verb{, }"model"\verb{, }"region")\verb{. Note that the }"variant"\verb{column referred to below contains information about the adaptation or variant name or}“N/A”`, as applicable.}
}
\value{
The output of \code{\link[=run_fredi]{run_fredi()}} is an R data frame object containing annual average impacts, by year (2010-2090), summarized at the specified aggregation and grouping levels.
}
\description{
Summarize and aggregate impacts from FrEDI (calculate national totals, average across models, sum impact types, and interpolate between impact estimate years).
}
\details{
This post-processing helper function aggregates and summarizes the FrEDI results to levels of aggregation specified by the user (passed to \code{aggLevels}). Users can specify a single aggregation level or multiple aggregation levels by passing a single character string or character vector to \code{aggLevels}. Options for aggregation include calculating national totals (\code{aggLevels="national"}), averaging across model types and models (\code{aggLevels="modelAverage"}), summing over all impact types (\code{aggLevels="impactType"}), and interpolating between impact year estimates (\code{aggLevels="impactYear"}). Users can specify all aggregation levels at once by specifying \code{aggLevels="all"} (default).
This post-processing helper function aggregates and summarizes temperature binning results to levels of aggregation specified by the user (passed to \code{aggLevels}). Users can specify a single aggregation level or multiple aggregation levels by passing a single character string or character vector to \code{aggLevels}. Options for aggregation include calculating national totals (\code{aggLevels="national"}), averaging across model types and models (\code{aggLevels="modelaverage"}), summing over all impact types (\code{aggLevels="impacttype"}), and interpolate between impact year estimates (\code{aggLevels="impactyear"}). Users can specify all aggregation levels at once by specifying \code{aggLevels="all"} (default) or no aggregation levels (\code{aggLevels="none"}).

Before aggregating impacts for national totals and/or model averages, \code{\link[=aggregate_impacts]{aggregate_impacts()}} will drop any pre-summarized results  (i.e., values for which \code{region="National Total"} and/or for which \code{model="average"}, respectively) that are already present in the data `and then reaggregate at those levels.

For each of the \code{aggLevels}, \code{\link[=aggregate_impacts]{aggregate_impacts()}} performs the following summarization (note that the \code{"variant"} column referred to below contains information about the adaptation or variant name or \verb{“N/A”}, as applicable):
\tabular{ll}{
\strong{Aggregation Level} \tab \strong{Description} \cr
\code{national} \tab Annual values are summed across all regions present in the data. I.e., data is grouped by columns \code{"sector"}, \code{"variant"}, \code{"impactType"},  \code{"impactYear"}, \code{"model_type"}, \code{"model"}, and \code{"year"}) and summed across regions. Years which have missing column data for all regions return as \code{NA}. The rows of the data frame of national values (with column \code{region="National Total"}) are then added as rows to the results. \cr
\code{modelaverage} \tab For temperature-driven sectors, annual results are averaged across all GCM models present in the data. I.e., data is grouped by columns \code{"sector"}, \code{"variant"}, \code{"impactType"}, \code{"impactYear"}, \code{"model_type"}, \code{"region"}, and \code{"year"} and averaged across models (SLR impacts are estimated as an interpolation between SLR scenarios). Averages exclude missing values. Years which have missing column data for all models return as \code{NA}. The rows of model averages (with column \code{model="Average"} are then added as rows to the results data frame. \cr
\code{impactType} \tab Annual results are summed across all impact types by sector present in the data. I.e., data is grouped by columns \code{"sector"}, \code{"variant"}, \code{"impactType"}, \code{"impactYear"},\code{"model_type"}, \code{"model"}, \code{"region"}, and \code{"year"} and summed across impact types. Mutates column \code{impactType="all"} for all values. Years which have missing column data for all impact types return as \code{NA}. If results are aggregated across impact types, information about physical impacts (columns \code{"physicalmeasure"} and \code{"physical_impacts"}) are dropped.\cr
\code{impactYear} \tab Annual results for sectors with only one impact year estimate (i.e., \code{impactYear == "N/A"}) are separated from those with multiple impact year estimates. For sectors with multiple impact years (i.e. 2010 and 2090 socioeconomic runs), annual results are interpolated between impact year estimates for applicable sectors  i.e., data is grouped by columns \verb{"sector", "variant", "impactType, "model_type", "model", "region", "year"} and interpolated across years with the 2010 run assigned to year 2010 and the 2090 run assigned to year 2090. The interpolated values are bound back to the results for sectors with a single impact year estimate, and column \code{impactYear} set to \code{impactYear="Interpolation"} for all values. \cr
}

Note that \code{\link[=aggregate_impacts]{aggregate_impacts()}} drops columns not used in grouping or aggregation.
}
\examples{
### Create temperature binning scenario
df_tempExOut <- run_fredi(aggLevels="none", pv=TRUE, silent=TRUE)

### Aggregate temperature binning summary across multiple columns
agg_tempExOut <- df_tempExOut \%>\% aggregate_impacts(columns=c("annual_impacts", "discounted_impacts"))

}
\references{
Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at \url{https://epa.gov/cira/FrEDI/}.
}
