% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcamScenarios}
\alias{gcamScenarios}
\title{Six driver scenarios that can be passed as inputs to \code{\link[=run_fredi]{run_fredi()}} and \code{\link[=run_fredi_sv]{run_fredi_sv()}}}
\format{
A data frame with 606 rows and 6 columns:
\describe{
\item{year         }{Year}
\item{temp_C_global}{Global temperatures, in degrees Celsius, for the associated year and scenario}
\item{temp_C_conus }{Temperatures for the contiguous U.S. (CONUS), in degrees Celsius, for the associated year and scenario (converted from global temperatures using \link{convertTemps(from="global")})}
\item{slr_cm       }{Global Mean Sea Level Rise (GMSL or SLR), in centimeters, for the associated year and scenario (calculated from global temperatures using \code{\link[=temps2slr]{temps2slr()}})}
\item{scenario     }{Associated scenario identifier (e.g., \code{"ECS_3.0_REF"})}
\item{model        }{A string (\code{"Hector_GCAM_v5.3"}) identifying the model (\strong{Hector}, with \strong{GCAM v5.3}) used in generating the global temperatures associated with each scenario}
}
}
\source{
\url{https://epa.gov/cira/FrEDI/}
}
\usage{
gcamScenarios
}
\description{
\code{gcamScenarios} is a data frame object containing six driver scenarios that can be passed as inputs to \code{\link[=run_fredi]{run_fredi()}} and \code{\link[=run_fredi_sv]{run_fredi_sv()}}. This data frame has five columns -- \code{year}, \code{temp_C_global}, \code{temp_C_conus}, \code{slr_cm}, \code{scenario}, and \code{model}:
}
\details{
\describe{
\item{year         }{The Common Era (CE) year associated with the observation.}
\item{temp_C_global}{Global temperature (i.e., degrees of warming above the baseline year of 1995), in degrees Celsius, for the associated year and scenario.}
\item{temp_C_conus }{Temperatures (i.e., degrees of warming above the baseline year of 1995) for the contiguous U.S. (CONUS), in degrees Celsius, for the associated year and scenario (calculated from global temperatures using \link{convertTemps(from="global")}).}
\item{slr_cm       }{Global Mean Sea Level Rise (GMSL or SLR), in centimeters, for the associated year and scenario (calculated from global temperatures using \code{\link[=temps2slr]{temps2slr()}}).}
\item{scenario     }{Associated scenario identifier (e.g., \code{"ECS_3.0_REF"}).}
\item{model        }{A string (\code{"Hector_GCAM_v5.3"}) identifying the model (\strong{Hector}, with \strong{GCAM v5.3}) used in generating the global temperatures associated with each scenario.}
}

The scenarios in this data frame were created using \strong{Hector} with \strong{GCAM v5.3}:
\itemize{
\item \href{https://jgcri.github.io/hector/}{\strong{Hector}} is an open-source, reduced-form model (Hartin et al., 2015) used to model temperatures associated with emissions scenarios .
\item \href{https://gcims.pnnl.gov/modeling/gcam-global-change-analysis-model}{\strong{GCAM v5.3}} -- an open source model that represents the linkages between physical and economic systems (Calvin et al., 2019).
}

Scenario identifiers in the \code{scenario} column of \code{\link[=gcamScenarios]{gcamScenarios()}} have the string \code{"ECS_3.0_REF_"} as a prefix, followed by a suffix indicating an emissions intensity associated with the scenario (e.g., \code{"20"}) -- for instance, the default scenario for \code{\link[=run_fredi]{run_fredi()}} and \code{\link[=run_fredi_sv]{run_fredi_sv()}} is \code{"ECS_3.0_REF"}. Other scenarios include \code{"ECS_3.0_REF_20"}, \code{"ECS_3.0_REF_30"}, \code{"ECS_3.0_REF_50"}, \code{"ECS_3.0_REF_70"}, and \code{"ECS_3.0_REF_90"}.

Users can use the scenarios in \code{gcamScenarios} as inputs to FrEDI or the FrEDI SV Module:
\itemize{
\item Users can filter the \code{gcamScenarios} data frame to any of these six scenarios, which can then be passed directly to the \code{\link[=run_fredi]{run_fredi()}} function via a named element (\code{temp} and/or \code{slr}) in a list passed to the \code{inputsList} argument -- e.g., \code{run_fredi(inputsList=list(temp=gcamScenarios |> filter(scenario=="ECS_3.0_REF")))} will run using the default temperature scenario.
\item Any or all of the GCAM scenarios can be passed directly to the FrEDI SV module via the \code{\link[=run_fredi_sv]{run_fredi_sv()}} function via a named list element (\code{temp} and/or \code{slr}) in a list passed to the \code{inputsList} argument -- e.g., \code{run_fredi_sv(inputsList=list(temp=gcamScenarios))} will run the SV module for all the GCAM scenarios provided in \code{gcamScenarios}.
\item \code{gcamScenarios} can also be combined with other provided scenarios (\code{gdpDefault}, \code{popDefault}) in function calls (e.g., \code{run_fredi(inputsList=list(temp=gcamScenarios |> filter(scenario=="ECS_3.0_REF"), gdp=gdpDefault, pop=popDefault))}) or user-provided data frames. For more information, see documentation for \code{\link[=run_fredi]{run_fredi()}}, \code{\link[=FrEDI:run_fredi_sv]{FrEDI:run_fredi_sv()}}, \code{\link[=FrEDI:run_fredi_methane]{FrEDI:run_fredi_methane()}}, and \code{\link[=import_inputs]{import_inputs()}}.
}

Calvin, K., Patel, P., Clarke, L., et al. 2019. \url{https://doi.org/10.5194/gmd-12-677-2019}.

Hartin, C.A., Patel, P., Schwarber, A., Link, R.P. and Bond-Lamberty, B.P., 2015. URL: <www.geosci-model-dev-discuss.net/7/7075/2014/>. \url{doi:10.5194/gmdd-7-7075-2014}.
}
\keyword{datasets}
