% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcamScenarios}
\alias{gcamScenarios}
\title{Six driver scenarios that can be passed as inputs to \code{\link[=run_fredi]{run_fredi()}} and \code{\link[=run_fredi_sv]{run_fredi_sv()}}}
\format{
A data frame with 546 rows and 4 columns:
\describe{
\item{year}{Year}
\item{temp_C}{Global temperature (in degrees Celsius) for the associated year and scenario}
\item{slr_cm}{Global Mean Sea Level Rise (in centimeters) for the associated year and scenario}
\item{scenario}{Associated scenario identifier}
}
}
\source{
\url{https://epa.gov/cira/FrEDI/}
}
\usage{
gcamScenarios
}
\description{
A dataframe containing six driver scenarios that can be passed as inputs to \code{\link[=run_fredi]{run_fredi()}} and \code{\link[=run_fredi_sv]{run_fredi_sv()}}. This data frame has four columns -- \code{year}, \code{temp_C}, \code{slr_cm}, and \code{scenario} -- respectively containing:
}
\details{
\describe{
\item{year}{The Common Era (CE) year associated with the observation.}
\item{temp_C}{Global temperature (i.e., degrees of warming above the baseline year of 1995) in degrees Celsius, for the associated year and scenario.}
\item{slr_cm}{Global Mean Sea Level Rise (i.e., SLR) in centimeters for the associated year and scenario.}
\item{scenario}{Associated scenario identifier.}
}

The scenarios in this dataframe were created using \href{https://jgcri.github.io/hector/}{Hector}, an open-source, reduced-form global carbon-cycle climate model (Hartin et al., 2015) to model temperatures associated with emissions scenarios from the Global Change Analysis Model v5.3 (GCAM). The Global Change Analysis Model v5.3 (\href{https://gcims.pnnl.gov/modeling/gcam-global-change-analysis-model}{GCAM}) is an open source model that represents the linkages between energy, water, land, climate and economic systems (Calvin et al., 2019). Scenario identifiers in the \code{scenario} column of \code{\link[=gcamScenarios]{gcamScenarios()}} have the string \code{"Hector_GCAM_v5.3_ECS_"} as a prefix, followed by the average warming temperature and the suffix \code{"_REF"} (e.g., \code{"Hector_GCAM_v5.3_ECS_3.0_REF"} for the default scenario for \code{\link[=run_fredi]{run_fredi()}} and \code{\link[=run_fredi_sv]{run_fredi_sv()}}.

These six temperature scenarios represent global temperatures; however, \code{\link[=run_fredi]{run_fredi()}} and \code{\link[=run_fredi_sv]{run_fredi_sv()}} require temperature scenarios for the contiguous U.S. (CONUS). Therefore, to use the \code{gcamScenarios} with \code{\link[=run_fredi]{run_fredi()}} or \code{\link[=run_fredi_sv]{run_fredi_sv()}}, users must first convert temperatures in the \code{temp_C} column to CONUS temperatures via the \link{convertTemps} function -- with argument \code{from = "global"} (e.g., \code{gcamScenarios |> mutate(temp_C = temp_C |> FrEDI::convertTemps(from = "global"))}).

The GCAM scenarios can be passed directly to the SV module via the \code{driverInput} argument (e.g., \code{run_fredi_sv(driverInput = gcamScenarios)}), since \code{\link[=run_fredi_sv]{run_fredi_sv()}} is designed to run multiple scenarios. In contrast, \code{\link[=run_fredi]{run_fredi()}} is intended to be run with a single scenario; \code{gcamScenarios} should be subset to a specific scenario before passing the temperature scenario to \code{\link[=run_fredi]{run_fredi()}} (e.g., \code{run_fredi(list(tempInput=gcamScenarios |> dplyr::filter(scenario=="Hector_GCAM_v5.3_ECS_3.0_REF")))} to use the default scenario).

Calvin, K., Patel, P., Clarke, L., et al. 2019. GCAM v5.1: representing the linkages between energy, water, land, climate, and economic systems, Geosci. Model Dev., 12:677–698. https://doi.org/10.5194/gmd-12-677-2019.

Hartin, C.A., Patel, P., Schwarber, A., Link, R.P. and Bond-Lamberty, B.P., 2015. A simple object-oriented and open-source model for scientific and policy analyses of the global climate system–Hector v1. 0. Geoscientific Model Development, 8(4), pp.939-955.
}
\keyword{datasets}
