% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcamScenarios}
\alias{gcamScenarios}
\title{Six driver scenarios that can be passed as inputs to \code{\link[=run_fredi_sv]{run_fredi_sv()}}}
\format{
A data frame with 546 rows and 4 columns:
\describe{
\item{year}{Year}
\item{temp_C}{CONUS temperature (in degrees Celsius) for the associated year and scenario}
\item{slr_cm}{Global Mean Sea Level Rise (in centimeters) for the associated year and scenario}
\item{scenario}{Associated scenario identifier}
}
}
\source{
\url{https://epa.gov/cira/FrEDI/}
}
\usage{
gcamScenarios
}
\description{
A dataframe containing six driver scenarios that can be passed as inputs to \code{\link[=run_fredi_sv]{run_fredi_sv()}}.
}
\details{
The scenarios in this dataframe were created using Hector, a reduced-form global carbon-cycle climate model (Hartin et al., 2015) to model temperatures associated with emissions scenarios from the Global Change Analysis Model v5.3 (GCAM). The Global Change Analysis Model v5.3 (GCAM) is an open source model that represents the linkages between energy, water, land, climate and economic systems (Calvin et al., 2019). For more information on Hector and GCAM, see Appendix C of the FrEDI documentation.

Calvin, K., Patel, P., Clarke, L., et al. 2019. GCAM v5.1: representing the linkages between energy, water, land, climate, and economic systems, Geosci. Model Dev., 12:677–698. https://doi.org/10.5194/gmd-12-677-2019.

Hartin, C.A., Patel, P., Schwarber, A., Link, R.P. and Bond-Lamberty, B.P., 2015. A simple object-oriented and open-source model for scientific and policy analyses of the global climate system–Hector v1. 0. Geoscientific Model Development, 8(4), pp.939-955.
}
\keyword{datasets}
