% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_inputs.R
\name{import_inputs}
\alias{import_inputs}
\title{Import custom scenarios for climate and socioeconomics (temperature and global mean sea level rise (GMSL), population, and GDP) from user-specified file names}
\usage{
import_inputs(
  tempfile = NULL,
  slrfile = NULL,
  popfile = NULL,
  gdpfile = NULL,
  temptype = "conus",
  popform = "wide"
)
}
\arguments{
\item{tempfile}{A character string indicating the location of a CSV file containing a custom temperature scenario (first column contains years in the interval 2000 to 2090 (values must start in 2000 or earlier); second column contains temperatures, in degrees Celsius, above the 1995 baseline).}

\item{slrfile}{A character string indicating the location of a CSV file containing a custom sea level rise scenario (first column contains years in the interval 2000 to 2090 (values must start in 2000 or earlier); second column contains values for global mean sea level rise (GMSL), in centimeters, above the 2000 baseline).}

\item{popfile}{A character string indicating the location of a CSV file containing a custom population scenario for NCA regions. The first column contains years in the interval 2010 to 2090. The number of additional columns, column names, and column contents depend on the population format set by \code{popform}. For more details, see \code{popform}.}

\item{gdpfile}{A character string indicating the location of a CSV file containing a custom scenario for gross domestic product (GDP) (first column contains years in the interval 2010 to 2090; second column contains values for GDP, in total 2015$).}

\item{temptype}{A character string indicating whether the temperature values in the temperature input file (specified by \code{tempfile} represent global temperature change (\code{temptype="global"}) or temperature change for the contiguous U.S. (\code{temptype="conus"}) in degrees Celsius. By default, the model assumes temperatures are CONUS temperatures (i.e., \code{default="conus"}).}

\item{popform}{A character string indicating whether the populations in the population input file specified by \code{popfile} are spread across multiple columns (i.e., \code{popform="wide"}) or are combined in a single column (i.e., \code{popform="long"}). By default, the model assumes \code{popform="wide"}. For both formats (\code{popform="wide"} or \code{popform="long"}), the first column contains values for the associated value year. If \code{popform="wide"} (default), the second through eighth columns of \code{popfile } must contain population values for each NCA region, with the associated NCA region as the column name. If \code{popform="long"}, the second column must contain NCA region names and the third column must contain values for the associated region population.}
}
\value{
\code{\link[=import_inputs]{import_inputs()}} returns a list of named elements containing dataframes with custom scenarios for temperature, GMSL, GDP, and regional population, respectively:

\tabular{ll}{
\strong{List Index} \tab \strong{Description} \cr
\code{tempInput} \tab Dataframe containing a custom temperature scenario imported from the CSV file specified by \code{tempfile}, with missing values removed. \code{tempInput} has two columns with names \verb{c("year"}, \verb{"temp_C")} containing the year and CONUS temperatures in degrees Celsius, respectively. \cr
\code{slrInput} \tab Dataframe containing a custom GMSL scenario imported from the CSV file specified by \code{slrfile}, with missing values removed. \code{slrInput} has two columns with names \verb{c("year"}, \verb{"slr_cm")} containing the year and global mean sea level rise (GMSL) in centimeters, respectively. \cr
\code{gdpInput} \tab Dataframe containing a custom GDP scenario imported from the CSV file specified by \code{gdpfile}, with missing values removed. \code{gdpInput} has two columns with names \code{c("year", "gdp_usd")} containing the year and the U.S. national GDP in 2015$, respectively. \cr
\code{popInput} \tab Dataframe containing a custom temperature scenario imported from the CSV file specified by \code{popfile}, with missing values removed. \code{popInput} has and three columns with names \verb{c("year"}, \code{"region"}, \verb{"reg_pop")} containing the year, the NCA region name, and the NCA region population, respectively. \cr
}
}
\description{
This function enables users to import data on custom scenarios for use with \link{FREDI} (supplied as inputs to \code{\link[=run_fredi]{run_fredi()}}). Users specify path names to CSV files containing temperature, global mean sea level rise (GMSL), gross domestic product (GDP), and population scenarios. \code{\link[=import_inputs]{import_inputs()}} reads in and format any specified files as data frames and returns a list of dataframes for imported scenarios.
}
\details{
This function enables users to import data on custom scenarios for use with temperature binning. Users specify path names to CSV files containing temperature, global mean sea level rise (GMSL), population, and gross domestic product (GDP) scenarios (\code{tempfile}, \code{slrfile}, \code{gdpfile}, and \code{popfile}, respectively). \code{\link[=import_inputs]{import_inputs()}} reads in and formats any specified files as data frames and returns a list of dataframes for imported scenarios. Users can specify whether the temperature input is for the contiguous U.S. (CONUS) or global using \code{temptype} and specify the format of the population scenario using \code{popform}.

Values for input scenarios must be within reasonable ranges. If a user inputs a custom scenario with values outside the allowable ranges, \code{\link[=import_inputs]{import_inputs()}} will not import the scenarios and will instead stop and return an error message. Temperature and GMSL inputs must begin in 2000 or earlier. Values for population and GDP scenarios can start in 2010 or earlier.
\itemize{
\item The input temperature scenario (passed to \code{\link[=run_fredi]{run_fredi()}} via the \code{inputsList} argument) requires temperatures for the contiguous U.S. (CONUS) in degrees Celsius relative to 1995 (degrees of warming relative to the baseline). Temperature values must be greater than or equal to zero and less than or equal to 10 degrees Celsius. Users can convert global temperatures to CONUS temperatures using \code{convertTemps(from="global")} or by specifying \code{import_inputs(temptype="global")} when importing a temperature scenario from a CSV file.
\item Values for the sea level rise (SLR) scenario are for global mean sea level rise (GMSL) must be in centimeters (cm) and values must be greater than or equal to zero and less than or equal to 250 cm.
\item Population and gross domestic product (GDP) values must be greater than or equal to zero.
}

If the population input is spread across multiple columns (i.e., \code{popform="wide"}), columns must be named according to the NCA regions. If the population input is in the long format, the region value must be in the second column. The NCA region names for population inputs must be in the following character vector: \verb{c("Midwest"}, \code{"Northeast"}, \code{"Northern.Plains"}, \code{"Northwest"}, \code{"Southeast"}, \code{"Southern.Plains"}, \verb{"Southwest")}. All regions must be present in the population input file.

\code{\link[=import_inputs]{import_inputs()}} outputs a list of dataframes that can be passed to the main FREDI function \code{\link[=run_fredi]{run_fredi()}} using the \code{inputList} argument. For example, specify \code{run_fredi(inputsList=x)} to generate impacts for a custom scenario \code{x} (where \code{x} is a list of dataframes such as that output from \code{\link[=import_inputs]{import_inputs()}}) (see \code{\link[=run_fredi]{run_fredi()}}).

All inputs to \code{\link[=import_inputs]{import_inputs()}} are optional. If the user does not specify a file path for \code{tempfile}, \code{slrfile}, \code{gdpfile}, or \code{popfile} (or if there is an error reading in inputs from those file paths), \code{\link[=import_inputs]{import_inputs()}} outputs a list with a \code{NULL} value for the associated list element. \code{\link[=run_fredi]{run_fredi()}} defaults back to the default scenarios for any elements of the inputs list that are \code{NULL} or missing. In other words, running \code{run_fredi(inputsList=list())} returns the same outputs as running \code{\link[=run_fredi]{run_fredi()}} (see \code{\link[=run_fredi]{run_fredi()}}).
}
\examples{
### Path to example scenarios
scenariosPath <- system.file(package="FrEDI") \%>\% file.path("extdata","scenarios")

### View example scenario names
scenariosPath \%>\% list.files

### Temperature Scenario File Name
tempInputFile <- scenariosPath \%>\% file.path("GCAM_scenario.csv")

### SLR Scenario File Name
slrInputFile  <- scenariosPath \%>\% file.path("slr_from_gcam.csv")

### Population Scenario File Name
popInputFile  <- scenariosPath \%>\% file.path("pop_scenario.csv")

### Import inputs
example_inputsList <- import_inputs(
  tempfile = tempInputFile,
  slrfile  = slrInputFile,
  popfile  = popInputFile,
  temptype = "global",
  popform  = "wide"
)

### Use imports with FREDI:
df_x <- run_fredi(inputsList=example_inputsList)

}
\references{
Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at \url{https://epa.gov/cira/FrEDI/}.
}
