% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fredi_sv.R
\name{run_fredi_sv}
\alias{run_fredi_sv}
\title{Calculate climate change impacts on socially vulnerable (SV) populations throughout the 21st century for available sectors}
\usage{
run_fredi_sv(
  sector = NULL,
  driverInput = NULL,
  popInput = NULL,
  silent = TRUE,
  save = FALSE,
  outpath = getwd(),
  overwrite = FALSE,
  addDate = FALSE,
  .testing = FALSE
)
}
\arguments{
\item{sector}{A character string indicating the sector for which the FrEDI SV module should calculate impacts (see \code{\link[=get_sv_sectorInfo]{get_sv_sectorInfo()}} for a list of available sectors).}

\item{driverInput}{A data frame of up to four custom scenarios for drivers (temperature or global mean sea level rise). \code{driverInput} requires a data frame with columns of \code{"year"} and \code{"scenario"}. The data frame must also include a third column: \code{"temp_C"} for temperature-driven sectors (containing temperature values in degrees Celsius of warming for the contiguous U.S.) or \code{"slr_cm"} for sea level rise (SLR)-driven sectors (containing values for global mean sea level rise in centimeters). Run \code{get_sv_sectorInfo(gcmOnly=TRUE)} to see temperature-driven sectors in the SV module and \code{get_sv_sectorInfo(slrOnly=TRUE)} to see SLR-driven scenarios. Users can also pass a data frame with all four columns (\code{"year"}, \code{"scenario"}, \code{"temp_C"}, and \code{"slr_cm"}), in which case \code{\link[=run_fredi_sv]{run_fredi_sv()}} determines whether to use the \code{"temp_C"} or \code{"slr_cm"} column as the driver trajectory based on the specified sector. Driver inputs for all scenarios should start in the year 2000 or earlier. All scenarios must include at least two non-missing values  (especially values before or at 2000 and at or after 2090). If any required columns are missing, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will use the default temperature or sea level rise scenario from \code{\link[=run_fredi]{run_fredi()}}. If the data frame passed to \code{driverInput} has more than four unique scenarios, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will only run the first four scenarios.}

\item{popInput}{A data frame containing regional population trajectories for each of the seven regions for the contiguous U.S. (\code{"Midwest"}, \code{"Northeast"}, \code{"Northern Plains"}, \code{"Northwest"}, \code{"Southeast"}, \code{"Southern Plains"}, \code{"Southwest"}) as defined by the \href{https://scenarios.globalchange.gov/regions_nca4}{National Climate Assessment (NCA)}. The data frame passed to \code{popInput} should have columns \code{"year"}, \code{"region"}, and \code{"reg_pop"}, which respectively contain values for year, NCA region name, and regional population. \code{popInput} only accepts a data frame with a single scenario; \code{\link[=run_fredi_sv]{run_fredi_sv()}} uses the same population scenario for any and all driver scenarios in the data frame passed to \code{driverInput}. If \code{popInput=NULL} (default), \code{\link[=run_fredi_sv]{run_fredi_sv()}} will use the default regional population trajectories.}

\item{silent}{A logical (\code{TRUE/FALSE}) value indicating the level of messaging desired by the user (defaults to \code{silent=TRUE}).}

\item{save}{A logical (\code{TRUE/FALSE}) value indicating whether to save the results to an Excel file (defaults to \code{save=FALSE}).}

\item{outpath}{A character string indicating a file directory to save the Excel file. Defaults to the working directory, i.e. \code{outpath=getwd()}. If the directory specified by \code{outpath} does not exist, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will attempt to create the specified directory.}

\item{overwrite}{A logical (\code{TRUE/FALSE}) value indicating whether to overwrite an existing Excel file if \code{save=TRUE} (defaults to \code{overwrite=FALSE}). If \code{overwrite=FALSE}, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will not automatically overwrite an existing Excel file; however, if a file exists and \code{overwrite=FALSE}, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will message the user and the user will have the option to overwrite the existing file. If \code{overwrite=TRUE} and the Excel file exists in the output directory, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will overwrite the existing file without messaging the user.}

\item{addDate}{A logical (\code{TRUE/FALSE}) value indicating whether to add the date to the name of the output Excel file if \code{save=TRUE} (defaults to \code{addDate=FALSE}). If \code{save=TRUE} and \code{addDate=TRUE}, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will append the system date to the beginning of the name of the outputs Excel file using the format \code{"\%Y\%m\%d"} (see \code{\link[base:format]{base::format()}} and \code{\link[base:Sys.time]{base::Sys.Date()}} for additional information).}
}
\value{
The output of \code{\link[=run_fredi_sv]{run_fredi_sv()}} is an R data frame object containing NCA region-specific annual average physical impacts for socially vulnerable groups at five-year increments between 2010 and 2090. An optional output of \code{\link[=run_fredi_sv]{run_fredi_sv()}} is an Excel file containing the output data frame with basic visualizations of sector outputs.
}
\description{
\code{run_fredi_sv} allows users to project annual average climate change impacts throughout the 21st century (2010-2090) for socially vulnerable (SV) populations for available sectors. Users can run \code{\link[=run_fredi_sv]{run_fredi_sv()}} for individual sectors to generate annual physical impacts for SV populations. \code{\link[=run_fredi_sv]{run_fredi_sv()}} can be run with default population and climate (temperature and sea level rise trajectories) or provide custom trajectories. The output of \code{\link[=run_fredi_sv]{run_fredi_sv()}} is an R data frame object containing annual average physical impacts at five-year increments for the period 2010 to 2090. Users have the option to write outputs to Excel files that provide additional visualization of SV outputs.
}
\details{
\code{\link[=run_fredi_sv]{run_fredi_sv()}} projects annual climate change impacts for socially vulnerable (SV) populations throughout the 21st century (2010-2090) for available sectors, using default or user-specified population, temperature, and sea level rise (SLR) trajectories. \code{\link[=run_fredi_sv]{run_fredi_sv()}} is the main function for the FrEDI Social Vulnerability (SV) module in the \link{FrEDI} R package, described elsewhere (See \url{https://epa.gov/cira/FrEDI} for more information). The SV module extends the \link{FrEDI} framework to socially vulnerable populations using data underlying a 2021 U.S. Environmental Protection Agency (EPA) report on \href{https://www.epa.gov/cira/social-vulnerability-report/}{Climate Change and Social Vulnerability in the United States}.

Users can run \code{\link[=run_fredi_sv]{run_fredi_sv()}} to generate annual physical impacts for SV groups for individual sectors. When running \code{\link[=run_fredi_sv]{run_fredi_sv()}}, users must specify one of the sectors in the SV module; use \code{\link[=get_sv_sectorInfo]{get_sv_sectorInfo()}} for a list of available sectors.

\code{\link[=run_fredi_sv]{run_fredi_sv()}} can be run with default population and climate (temperature and SLR) trajectories or use \code{\link[=run_fredi_sv]{run_fredi_sv()}} to run custom scenarios. Running \code{\link[=run_fredi_sv]{run_fredi_sv()}} with custom climate scenarios requires passing a data frame of scenarios to the \code{driverInput} argument. \code{\link[=run_fredi_sv]{run_fredi_sv()}} can also be run with a custom population scenario by passing a data frame of regional population trajectories to the \code{popInput} argument; unlike climate scenarios, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will only run a single scenario at a time.
\itemize{
\item \code{driverInput} can take a data frame containing up to four custom scenarios for drivers (temperature or global mean sea level rise). \code{driverInput} requires a data frame with columns of \code{"year"} and \code{"scenario"}. The data frame must also include a third column: \code{"temp_C"} for temperature-driven sectors (containing temperature values in degrees Celsius of warming for the contiguous U.S.) or \code{"slr_cm"} for sea level rise (SLR)-driven sectors (containing values for global mean sea level rise in centimeters). Run \code{get_sv_sectorInfo(gcmOnly=TRUE)} to see temperature-driven sectors in the SV module and \code{get_sv_sectorInfo(slrOnly=TRUE)} to see SLR-driven scenarios. Users can also pass a data frame with all four columns (\code{"year"}, \code{"scenario"}, \code{"temp_C"}, and \code{"slr_cm"}), in which case \code{\link[=run_fredi_sv]{run_fredi_sv()}} determines whether to use the \code{"temp_C"} or \code{"slr_cm"} column as the driver trajectory based on the specified sector. If any required columns are missing, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will use the default temperature or sea level rise scenario from \code{\link[=run_fredi]{run_fredi()}}. If the data frame passed to \code{driverInput} has more than four unique scenarios, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will only run the first four scenarios.
\itemize{
\item Temperature inputs must be temperature change in degrees Celsius for the contiguous U.S. (use \code{\link[=convertTemps]{convertTemps()}} to convert global temperatures to CONUS temperatures before passing to \code{driverInput}) relative to a 1995 baseline (where 1995 is the central year of a 1986-2005 baseline period; values should start at zero in the year 1995).
\item Sea level rise inputs must be in centimeters relative to a 2000 baseline (i.e., values should start at zero in the year 2000). Driver inputs for all scenarios should start in the year 2000 or earlier. All scenarios must include at least two non-missing values  (especially values before or at 2000 and at or after 2090).
}
\item \code{popInput} can take a data frame containing a single scenario with regional population trajectories for each of the seven regions for the contiguous U.S. (\code{"Midwest"}, \code{"Northeast"}, \code{"Northern Plains"}, \code{"Northwest"}, \code{"Southeast"}, \code{"Southern Plains"}, \code{"Southwest"}) as defined by the \href{https://scenarios.globalchange.gov/regions_nca4}{National Climate Assessment (NCA)}. The data frame passed to \code{popInput} should have columns \code{"year"}, \code{"region"}, and \code{"reg_pop"}, which respectively contain values for year, NCA region name, and regional population. \code{popInput} only accepts a data frame with a single scenario; \code{\link[=run_fredi_sv]{run_fredi_sv()}} uses the same population scenario for any and all driver scenarios in the data frame passed to \code{driverInput}. If \code{popInput=NULL} (default), \code{\link[=run_fredi_sv]{run_fredi_sv()}} will use the default regional population trajectories. The default regional population scenario is drawn from the Integrated Climate and Land Use Scenarios version 2 (ICLUSv2) model (Bierwagen et al, 2010; EPA 2017) under the Median variant projection of United Nations (United Nations, 2015). Note that the FrEDI SV default population scenario differs from the default population scenario used by \code{\link[=run_fredi]{run_fredi()}}.
}

The output of \code{\link[=run_fredi_sv]{run_fredi_sv()}} is an R data frame object containing NCA region-specific annual average physical impacts for socially vulnerable groups at five-year increments between 2010 and 2090. Users have the additional option to write results to an Excel file by setting \code{save=TRUE}; output Excel files provide basic visualizations of output data. Additional arguments provide more control over how the outputs are saved if \code{save=TRUE}:
\itemize{
\item \code{outpath} can be used to specify the directory in which to save an Excel output file. Defaults to \code{outpath=getwd()} (i.e., the working directory). If the directory specified by \code{outpath} does not exist, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will attempt to create the specified directory.
\item \code{overwrite} can be used to force \code{\link[=run_fredi_sv]{run_fredi_sv()}} to overwrite an existing Excel file in the output directory.  If \code{overwrite=FALSE}, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will not automatically overwrite an existing Excel file; however, if a file exists and \code{overwrite=FALSE}, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will message the user and the user will have the option to overwrite the existing file. If \code{overwrite=TRUE} and the Excel file exists in the output directory, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will overwrite the existing file without messaging the user.
\item \code{addDate} can be used to append the date to the output Excel file. If \code{save=TRUE} and \code{addDate=TRUE}, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will append the system date to the beginning of the name of the outputs Excel file using the format \code{"\%Y\%m\%d"} (see \code{\link[base:format]{base::format()}} and \code{\link[base:Sys.time]{base::Sys.Date()}} for additional information).
}
}
\examples{
### Run SV Module with defaults without specifying sector
df_sv <- run_fredi_sv()

### Return a character vector with the names of all of the sectors in the FrEDI SV Module:
get_sv_sectorInfo()

### Return a data frame of all of the sectors in the FrEDI SV Module (sector names and additional information)
get_sv_sectorInfo(description=T)

### Run SV Module with defaults for "Coastal Properties" without saving
df_sv <- run_fredi_sv(sector="Coastal Properties")

### Run SV Module with defaults for "Extreme Temperature" without saving
df_sv <- run_fredi_sv(sector="Extreme Temperature")

### Run SV Module with defaults for "Extreme Temperature" with saving and add date to file name
df_sv <- run_fredi_sv(sector="Extreme Temperature", save=T, addDate=T)

### Load temperature scenarios
load(gcamScenarios)

### Load population scenario
load(popScenario)

### Run SV Module for "Extreme Temperature" with custom population and temperature scenarios. Save and overwrite previous results
df_sv <- run_fredi_sv(sector="Extreme Temperature", driverInput = gcamScenarios, popInput = popScenario, save=T, addDate=T, overwrite = T)

}
\references{
Bierwagen, B., D. M. Theobald, C. R. Pyke, A. Choate, P. Groth, J. V. Thomas, and P. Morefield. 2010. “National housing and impervious surface scenarios for integrated climate impact assessments.” Proc. Natl. Acad. Sci. 107 (49): 20887–20892. https://doi.org/10.1073/pnas.1002096107.

EPA. 2017. Multi-Model Framework for Quantitative Sectoral Impacts Analysis: A technical report for the Fourth National Climate Assessment. U.S. Environmental Protection Agency, EPA 430-R-17-001.

EPA. 2021. Technical Documentation on the Framework for Evaluating Damages and Impacts (FrEDI). U.S. Environmental Protection Agency, EPA 430-R-21-004. Available at \url{https://epa.gov/cira/FrEDI/}.

EPA. 2021. Climate Change and Social Vulnerability in the United States: A Focus on Six Impacts. U.S. Environmental Protection Agency, EPA 430-R-21-003. Available at \url{https://www.epa.gov/cira/social-vulnerability-report/}.

United Nations. 2015. World population prospects: The 2015 revision. New York: United Nations, Department of Economic and Social Affairs, Population Division.

U.S. Global Change Research Program. 2015. Scenarios for the National Climate Assessment. Available at \url{https://scenarios.globalchange.gov/regions_nca4}.
}
