% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fredi_sv.R
\name{run_fredi_sv}
\alias{run_fredi_sv}
\title{FrEDI Social Vulnerability (SV) Module: Calculates climate change impacts on socially vulnerable populations throughout the 21st century for available sectors}
\usage{
run_fredi_sv(
  inputsList = list(driverInput = NULL, popInput = NULL),
  sector = NULL,
  silent = TRUE,
  save = FALSE,
  outpath = getwd(),
  overwrite = FALSE
)
}
\arguments{
\item{inputsList}{A list of named elements named elements (\code{names(inputsList)= c("driverInput",  "popInput")}), each containing dataframes of custom scenarios for drivers (temperature or global mean sea level rise) and/or regional population scenarios, respectively. Scenarios must be defined over the period 2010 to 2090. Note: driver inputs (temperature and sea level rise) should start in 2000 or earlier. Population inputs should start in 2010 or earlier. Temperature inputs must be in degrees Celsius for the CONUS region. If starting from global temperature change, first use \code{\link[=convertTemps]{convertTemps()}} to convert global temperatures to CONUS temperatures.}

\item{sector=NULL}{A character string indicating for which sector to run the FrEDI SV module.}

\item{silent=TRUE}{A \code{TRUE/FALSE} value indicating the level of messaging desired by the user (default=\code{TRUE}).}

\item{save=FALSE}{A \code{TRUE/FALSE} value indicating whether to save the results to an Excel file (default=\code{FALSE}).}

\item{outpath=getwd()}{A character string indicating a file directory to save the Excel file (created in \code{save=TRUE}). By default, if \code{save=TRUE}, the Excel file will be saved to the working directory (with a file name determined by the sector).}

\item{overwrite=FALSE}{A \code{TRUE/FALSE} value indicating whether to overwrite an existing Excel file if \code{save=TRUE} (default=\code{FALSE}). By default, if \code{save=TRUE}, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will not write over an existing file.}
}
\value{
The output of \code{\link[=run_fredi_sv]{run_fredi_sv()}} is an R data frame object containing annual average impacts, by year (2010-2090), for each sector, adaptation, model (GCM or SLR scenario), and region.
}
\description{
This function allows users to project annual average climate change impacts throughout the 21st century (2010-2090) for socially vulnerable populations for available sectors (see \code{\link[=get_sectorInfo]{get_sectorInfo()}}). Users may specify an optional list of custom scenarios. The output is an R data frame object containing annual average impacts, by year.
}
\details{
This function allows users to project annual average climate change impacts throughout the 21st century (2010-2090) for available sectors. \code{\link[=run_fredi_sv]{run_fredi_sv()}} is the main function in the \link{FrEDI} R package, described elsewhere (See \url{https://epa.gov/cira/FrEDI} for more information).
}
\examples{
### Run function with defaults (same as `defaultResults` dataset)
sv_defaults <- run_fredi_sv()

### Path to example scenarios
scenariosPath <- system.file(package="FrEDI") \%>\% file.path("extdata","scenarios")
### View example scenario names
scenariosPath \%>\% list.files
### Temperature Scenario File Name
tempInputFile <- scenariosPath \%>\% file.path("GCAM_scenario.csv")
### SLR Scenario File Name
slrInputFile  <- scenariosPath \%>\% file.path("slr_from_GCAM.csv")
### Population Scenario File Name
popInputFile  <- scenariosPath \%>\% file.path("pop_scenario.csv")
### Import inputs
example_inputsList <- import_inputs(
  tempfile = tempInputFile,
  slrfile  = slrInputFile,
  popfile  = popInputFile
)

### Run custom temperature scenario and output impacts without aggregation and with present values (default base year and discount rate)
df_tempExOut <- run_fredi(inputsList= tempBin_inputs, aggLevels="none", pv=TRUE, silent=TRUE)

}
\references{
Environmental Protection Agency (EPA). 2021. Technical Documentation on The Framework for Evaluating Damages and Impacts (FrEDI). Technical Report EPA 430-R-21-004, EPA, Washington, DC. Available at \url{https://epa.gov/cira/FrEDI/}.
}
